requirements.txt
pyautogui
keyboard
pyperclip
python-dotenv